Treat POH deaths as safe
This plugin considers a death to be unsafe if it respawns in a list of respawn
locations. If your POH is in prifddinas and you die in it, however, you will
respawn in the same chunk as the prifddinas unsafe death respawn point, so
previously the plugin thinks that you died an unsafe death. Change the behavior
to check if the player died in a POH and treat it as a safe death instead.